Search filter
I added a search filter to search through the notes with js. I have also been practicing utilizing detailed comments, so that I can refer back to previous work for examples of how to solve problems in the future. I still need to touch up the styling, but apart from that this is finished.